+++ lib/src/entities/image_picker_display.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:image_picker_plus/image_picker_plus.dart';
 import 'package:flutter/material.dart';
 
@@ -7,6 +8,7 @@ class GalleryDisplaySettings {
   TabsTexts? tabsTexts;
   SliverGridDelegateWithFixedCrossAxisCount gridDelegate;
   bool showImagePreview;
+  AsyncValueSetter<SelectedImagesDetails>? sendRequestFunction;
 
   /// If [cropImage] true [showImagePreview] will be true
   /// Right now this package not support crop video
@@ -14,6 +16,7 @@ class GalleryDisplaySettings {
 
   GalleryDisplaySettings({
     this.appTheme,
+    this.sendRequestFunction,
     this.tabsTexts,
     this.gridDelegate = const SliverGridDelegateWithFixedCrossAxisCount(
         crossAxisCount: 4, crossAxisSpacing: 1.7, mainAxisSpacing: 1.5),