+++ lib/src/image_picker_plus.dart
@@ -52,14 +52,15 @@ class ImagePickerPlus {
     bool multiSelection = false,
     required PickerSource pickerSource,
   }) async {
-    return await Navigator.of(_context).push(
+    return await Navigator.of(_context,rootNavigator: true).push(
       MaterialPageRoute(
-        builder: (context) => CustomImagePicker(
+        builder: (context,) => CustomImagePicker(
           galleryDisplaySettings: galleryDisplaySettings,
           multiSelection: multiSelection,
           pickerSource: pickerSource,
           source: source,
         ),
+        maintainState: false,
       ),
     );
   }